Move cmdline net tests into test_net_cmdline file
