+++ tests/unittests/test_net.py
@@ -4016,237 +4016,6 @@ class TestEniNetworkStateToEni(CiTestCase):
         self.assertNotIn("hwaddress", rendered)
 
 
-class TestCmdlineConfigParsing(CiTestCase):
-    with_logs = True
-
-    simple_cfg = {
-        'config': [{"type": "physical", "name": "eth0",
-                    "mac_address": "c0:d6:9f:2c:e8:80",
-                    "subnets": [{"type": "dhcp"}]}]}
-
-    def test_cmdline_convert_dhcp(self):
-        found = cmdline._klibc_to_config_entry(DHCP_CONTENT_1)
-        self.assertEqual(found, ('eth0', DHCP_EXPECTED_1))
-
-    def test_cmdline_convert_dhcp6(self):
-        found = cmdline._klibc_to_config_entry(DHCP6_CONTENT_1)
-        self.assertEqual(found, ('eno1', DHCP6_EXPECTED_1))
-
-    def test_cmdline_convert_static(self):
-        found = cmdline._klibc_to_config_entry(STATIC_CONTENT_1)
-        self.assertEqual(found, ('eth1', STATIC_EXPECTED_1))
-
-    def test_config_from_cmdline_net_cfg(self):
-        files = []
-        pairs = (('net-eth0.cfg', DHCP_CONTENT_1),
-                 ('net-eth1.cfg', STATIC_CONTENT_1))
-
-        macs = {'eth1': 'b8:ae:ed:75:ff:2b',
-                'eth0': 'b8:ae:ed:75:ff:2a'}
-
-        dhcp = copy.deepcopy(DHCP_EXPECTED_1)
-        dhcp['mac_address'] = macs['eth0']
-
-        static = copy.deepcopy(STATIC_EXPECTED_1)
-        static['mac_address'] = macs['eth1']
-
-        expected = {'version': 1, 'config': [dhcp, static]}
-        with temp_utils.tempdir() as tmpd:
-            for fname, content in pairs:
-                fp = os.path.join(tmpd, fname)
-                files.append(fp)
-                util.write_file(fp, content)
-
-            found = cmdline.config_from_klibc_net_cfg(files=files,
-                                                      mac_addrs=macs)
-            self.assertEqual(found, expected)
-
-    def test_cmdline_with_b64(self):
-        data = base64.b64encode(json.dumps(self.simple_cfg).encode())
-        encoded_text = data.decode()
-        raw_cmdline = 'ro network-config=' + encoded_text + ' root=foo'
-        found = cmdline.read_kernel_cmdline_config(cmdline=raw_cmdline)
-        self.assertEqual(found, self.simple_cfg)
-
-    def test_cmdline_with_net_config_disabled(self):
-        raw_cmdline = 'ro network-config=disabled root=foo'
-        found = cmdline.read_kernel_cmdline_config(cmdline=raw_cmdline)
-        self.assertEqual(found, {'config': 'disabled'})
-
-    def test_cmdline_with_net_config_unencoded_logs_error(self):
-        """network-config cannot be unencoded besides 'disabled'."""
-        raw_cmdline = 'ro network-config={config:disabled} root=foo'
-        found = cmdline.read_kernel_cmdline_config(cmdline=raw_cmdline)
-        self.assertIsNone(found)
-        expected_log = (
-            'ERROR: Expected base64 encoded kernel commandline parameter'
-            ' network-config. Ignoring network-config={config:disabled}.')
-        self.assertIn(expected_log, self.logs.getvalue())
-
-    def test_cmdline_with_b64_gz(self):
-        data = _gzip_data(json.dumps(self.simple_cfg).encode())
-        encoded_text = base64.b64encode(data).decode()
-        raw_cmdline = 'ro network-config=' + encoded_text + ' root=foo'
-        found = cmdline.read_kernel_cmdline_config(cmdline=raw_cmdline)
-        self.assertEqual(found, self.simple_cfg)
-
-
-class TestCmdlineKlibcNetworkConfigSource(FilesystemMockingTestCase):
-    macs = {
-        'eth0': '14:02:ec:42:48:00',
-        'eno1': '14:02:ec:42:48:01',
-    }
-
-    def test_without_ip(self):
-        content = {'/run/net-eth0.conf': DHCP_CONTENT_1,
-                   cmdline._OPEN_ISCSI_INTERFACE_FILE: "eth0\n"}
-        exp1 = copy.deepcopy(DHCP_EXPECTED_1)
-        exp1['mac_address'] = self.macs['eth0']
-
-        root = self.tmp_dir()
-        populate_dir(root, content)
-        self.reRoot(root)
-
-        src = cmdline.KlibcNetworkConfigSource(
-            _cmdline='foo root=/root/bar', _mac_addrs=self.macs,
-        )
-        self.assertTrue(src.is_applicable())
-        found = src.render_config()
-        self.assertEqual(found['version'], 1)
-        self.assertEqual(found['config'], [exp1])
-
-    def test_with_ip(self):
-        content = {'/run/net-eth0.conf': DHCP_CONTENT_1}
-        exp1 = copy.deepcopy(DHCP_EXPECTED_1)
-        exp1['mac_address'] = self.macs['eth0']
-
-        root = self.tmp_dir()
-        populate_dir(root, content)
-        self.reRoot(root)
-
-        src = cmdline.KlibcNetworkConfigSource(
-            _cmdline='foo ip=dhcp', _mac_addrs=self.macs,
-        )
-        self.assertTrue(src.is_applicable())
-        found = src.render_config()
-        self.assertEqual(found['version'], 1)
-        self.assertEqual(found['config'], [exp1])
-
-    def test_with_ip6(self):
-        content = {'/run/net6-eno1.conf': DHCP6_CONTENT_1}
-        root = self.tmp_dir()
-        populate_dir(root, content)
-        self.reRoot(root)
-
-        src = cmdline.KlibcNetworkConfigSource(
-            _cmdline='foo ip6=dhcp root=/dev/sda', _mac_addrs=self.macs,
-        )
-        self.assertTrue(src.is_applicable())
-        found = src.render_config()
-        self.assertEqual(
-            found,
-            {'version': 1, 'config': [
-             {'type': 'physical', 'name': 'eno1',
-              'mac_address': self.macs['eno1'],
-              'subnets': [
-                  {'dns_nameservers': ['2001:67c:1562:8010::2:1'],
-                   'control': 'manual', 'type': 'dhcp6', 'netmask': '64'}]}]})
-
-    def test_with_no_ip_or_ip6(self):
-        # if there is no ip= or ip6= on cmdline, return value should be None
-        content = {'net6-eno1.conf': DHCP6_CONTENT_1}
-        files = sorted(populate_dir(self.tmp_dir(), content))
-        src = cmdline.KlibcNetworkConfigSource(
-            _files=files, _cmdline='foo root=/dev/sda', _mac_addrs=self.macs,
-        )
-        self.assertFalse(src.is_applicable())
-
-    def test_with_both_ip_ip6(self):
-        content = {
-            '/run/net-eth0.conf': DHCP_CONTENT_1,
-            '/run/net6-eth0.conf': DHCP6_CONTENT_1.replace('eno1', 'eth0')}
-        eth0 = copy.deepcopy(DHCP_EXPECTED_1)
-        eth0['mac_address'] = self.macs['eth0']
-        eth0['subnets'].append(
-            {'control': 'manual', 'type': 'dhcp6',
-             'netmask': '64', 'dns_nameservers': ['2001:67c:1562:8010::2:1']})
-        expected = [eth0]
-
-        root = self.tmp_dir()
-        populate_dir(root, content)
-        self.reRoot(root)
-
-        src = cmdline.KlibcNetworkConfigSource(
-            _cmdline='foo ip=dhcp ip6=dhcp', _mac_addrs=self.macs,
-        )
-
-        self.assertTrue(src.is_applicable())
-        found = src.render_config()
-        self.assertEqual(found['version'], 1)
-        self.assertEqual(found['config'], expected)
-
-
-class TestReadInitramfsConfig(CiTestCase):
-
-    def _config_source_cls_mock(self, is_applicable, render_config=None):
-        return lambda: mock.Mock(
-            is_applicable=lambda: is_applicable,
-            render_config=lambda: render_config,
-        )
-
-    def test_no_sources(self):
-        with mock.patch('cloudinit.net.cmdline._INITRAMFS_CONFIG_SOURCES', []):
-            self.assertIsNone(cmdline.read_initramfs_config())
-
-    def test_no_applicable_sources(self):
-        sources = [
-            self._config_source_cls_mock(is_applicable=False),
-            self._config_source_cls_mock(is_applicable=False),
-            self._config_source_cls_mock(is_applicable=False),
-        ]
-        with mock.patch('cloudinit.net.cmdline._INITRAMFS_CONFIG_SOURCES',
-                        sources):
-            self.assertIsNone(cmdline.read_initramfs_config())
-
-    def test_one_applicable_source(self):
-        expected_config = object()
-        sources = [
-            self._config_source_cls_mock(
-                is_applicable=True, render_config=expected_config,
-            ),
-        ]
-        with mock.patch('cloudinit.net.cmdline._INITRAMFS_CONFIG_SOURCES',
-                        sources):
-            self.assertEqual(expected_config, cmdline.read_initramfs_config())
-
-    def test_one_applicable_source_after_inapplicable_sources(self):
-        expected_config = object()
-        sources = [
-            self._config_source_cls_mock(is_applicable=False),
-            self._config_source_cls_mock(is_applicable=False),
-            self._config_source_cls_mock(
-                is_applicable=True, render_config=expected_config,
-            ),
-        ]
-        with mock.patch('cloudinit.net.cmdline._INITRAMFS_CONFIG_SOURCES',
-                        sources):
-            self.assertEqual(expected_config, cmdline.read_initramfs_config())
-
-    def test_first_applicable_source_is_used(self):
-        first_config, second_config = object(), object()
-        sources = [
-            self._config_source_cls_mock(
-                is_applicable=True, render_config=first_config,
-            ),
-            self._config_source_cls_mock(
-                is_applicable=True, render_config=second_config,
-            ),
-        ]
-        with mock.patch('cloudinit.net.cmdline._INITRAMFS_CONFIG_SOURCES',
-                        sources):
-            self.assertEqual(first_config, cmdline.read_initramfs_config())
-
-
 class TestNetplanRoundTrip(CiTestCase):
 
     NETPLAN_INFO_OUT = textwrap.dedent("""
@@ -5139,14 +4908,6 @@ class TestGetIBHwaddrsByInterface(CiTestCase):
         result = net.get_ib_hwaddrs_by_interface()
         expected = {'ib0': self._ib_addr}
         self.assertEqual(expected, result)
-
-
-def _gzip_data(data):
-    with io.BytesIO() as iobuf:
-        gzfp = gzip.GzipFile(mode="wb", fileobj=iobuf)
-        gzfp.write(data)
-        gzfp.close()
-        return iobuf.getvalue()
 
 
 class TestRenameInterfaces(CiTestCase):
